Move rebroadcast implementation
In preparation for dataplane TCP support
