+++ monad-raptorcast/benches/raptor_bench.rs
@@ -8,9 +8,8 @@ use monad_crypto::hasher::{Hasher, HasherType};
 use monad_dataplane::network::MONAD_GSO_SIZE;
 use monad_raptor::ManagedDecoder;
 use monad_raptorcast::{
-    udp::{build_messages, parse_message},
+    udp::{build_messages, parse_message, SIGNATURE_CACHE_SIZE},
     util::{BuildTarget, EpochValidators, Validator},
-    SIGNATURE_CACHE_SIZE,
 };
 use monad_secp::{KeyPair, SecpSignature};
 use monad_types::{NodeId, Stake};
